As I understand it, from the Java Version I looked at, This is basically two structures: A hash table for a fast lookup, and a doubly linked list.  Every time you "hit" something, you move it to the top of the list.
I want to keep this generic, but the value must be sized.  The
eviction routine is based on the maximum amount of memory the
value can be.

I really don't feel like I know what I'm doing here.  This doesn't
seem to build on itself correctly, although it works on paper!
Grr.  I want values to MOVE into the HashMap and LIVE there, and I
want them to go away when the value is deleted, and I want
REFERENCES to the value to be made available to consumers
(i.e. consumers may borrow the value, but immutably and
non-permanently.)