Work around Hy's scoping limitations and now it passes the tests.
After months of writing Hy and never hitting the so-called "let/scoping"
issue that plagues the Hy team, and getting really comfortable with Racket,
I finally hit a place where my Lisp skills and my Python skills collided
to prevent the code from working: to wit, deeper variable assignments
shadowed earlier ones in ways Lisp tolerates and Python doesn't.  Re-writing
the entire server to use setv and class-based scoping really brings home
just how badly Hy Is Just Python messes with my head.

I do love Hy's syntax and a lot of the expression-oriented code that's
part of Hy's (and Coffeescript's) Lisp legacy, but this is one of those
places where it drives me nuts.

The server now works more or less as expected.  It is NOT correct, not
even a little bit; there are aspects to it that break easily, but as a
learning testbed for how this works, it's Good Enough and now I have an
excuse to move on to the Rust version.